Samples: nrf5340: Ensure SDC asserts are printed before reset
Log subsystem is not in panic mode when BT_ASSERT_MSG is called for some
reason, and without it, there is a chance that the NET core is reset
before having a chance to flush it's UART buffer.

This commit enables panic mode before logging the controller assert in
the multiprotocol_rpmsg sample.

Signed-off-by: Jonathan Rico <jonathan.rico@nordicsemi.no>